move functions to token utils
