+++ lib/parser.js
@@ -1,4 +1,7 @@
 var tokenizer = require('./tokenizer');
+
+var tokenUtils = require('./util/tokenUtils');
+
 var classHandlers = require('./handlers/classHandlers');
 var functionHandlers = require('./handlers/functionHandlers');
 var moduleHandlers = require('./handlers/moduleHandlers');
@@ -116,63 +119,6 @@ var valid = [
     'git',
     'sit'
 ];
-
-/**
- * Raises an error if the given tokens do not start with the desired state
- */
-function expectToken(expectedStart, parseContext)
-{
-  var tokens = parseContext.tokens;
-  if ( tokens[0] !== expectedStart)
-  {
-    throw new Error(`Invalid parse state! Expected: '${expectedStart}' but got: '${tokens[0]}' from chain: [${tokens}]. ${parseInfo(parseContext)}`);
-  }
-}
-
-/**
- * Raises an error if the start token does not match any of the expected tokens.
- */
-function expectAnyToken(expectedTokens, parseContext)
-{
-  var tokens = parseContext.tokens;
-
-  var foundToken = false;
-  var firstToken = tokens[0];
-  for (let i = 0; i < expectedTokens.length; i ++)
-  {
-    if(firstToken === expectedTokens[i])
-    {
-      foundToken = true;
-      break;
-    }
-  }
-
-  if(!foundToken)
-  {
-    throw new Error(`Invalid parse state! Expected any of: '${expectedTokens}' but got: '${tokens[0]}' from chain: [${tokens}]. ${parseInfo(parseContext)}`);
-  }
-}
-
-/**
- * Joins the set of tokens into a string such that joinTokens([a, b]) returns "a b".
- * Consumes all tokens from the given array such that the array will be empty if it is referenced after calling this function.
- */
-function joinTokens(tokens)
-{
-  var tokenString = '';
-
-  let token;
-  while(token = tokens.shift())
-  {
-    tokenString += token;
-    if(tokens.length > 0)
-    {
-      tokenString += ' ';
-    }
-  }
-
-  return tokenString;
-}
 
 function containsUnary(parseContext)
 {
@@ -330,7 +276,7 @@ function isAssignmentOperator(token)
  */
 function handleWoof(parseContext)
 {
-  expectToken('woof', parseContext);
+  tokenUtils.expectToken('woof', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -392,7 +338,7 @@ function handleWoof(parseContext)
  */
 function handleShh(parseContext)
 {
-  expectToken('shh', parseContext);
+  tokenUtils.expectToken('shh', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -422,7 +368,7 @@ function handleShh(parseContext)
  */
  function handleQuiet(parseContext)
  {
-  expectToken('quiet', parseContext);
+  tokenUtils.expectToken('quiet', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -432,7 +378,7 @@ function handleShh(parseContext)
    multiComment = true;
 
    var statement = '/*';
-   statement += joinTokens(parseContext.tokens);
+   statement += tokenUtils.joinTokens(parseContext.tokens);
    statement += '\n';
 
    return statement;
@@ -446,7 +392,7 @@ function handleShh(parseContext)
  */
 function handleLoud(parseContext)
 {
-  expectToken('loud', parseContext);
+  tokenUtils.expectToken('loud', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -461,7 +407,7 @@ function handleLoud(parseContext)
   multiComment = false;
 
   var statement = '*/';
-  statement += joinTokens(parseContext.tokens);
+  statement += tokenUtils.joinTokens(parseContext.tokens);
   statement += '\n';
 
   return statement;
@@ -475,7 +421,7 @@ function handleLoud(parseContext)
  */
 function handleNew(parseContext)
 {
-  expectToken('new', parseContext);
+  tokenUtils.expectToken('new', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -512,7 +458,7 @@ function handleNew(parseContext)
  */
 function handleRly(parseContext)
 {
-  expectToken('rly', parseContext);
+  tokenUtils.expectToken('rly', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -534,7 +480,7 @@ function handleRly(parseContext)
  */
 function handleNotrly(parseContext)
 {
-  expectToken('notrly', parseContext);
+  tokenUtils.expectToken('notrly', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -557,7 +503,7 @@ function handleNotrly(parseContext)
  */
 function handleBut(parseContext)
 {
-  expectToken('but', parseContext);
+  tokenUtils.expectToken('but', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -592,7 +538,7 @@ function handleBut(parseContext)
  */
 function handleWith(parseContext)
 {
-  expectToken('with', parseContext);
+  tokenUtils.expectToken('with', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -687,7 +633,7 @@ function handleWith(parseContext)
  */
 function handleWow(parseContext)
 {
-  expectAnyToken(['wow', 'wow&'], parseContext);
+  tokenUtils.expectAnyToken(['wow', 'wow&'], parseContext);
 
   var tokens = parseContext.tokens;
   multiLine = false;
@@ -733,7 +679,7 @@ function handleWow(parseContext)
  */
 function handleMuchLoop(parseContext)
 {
-  expectToken('much', parseContext);
+  tokenUtils.expectToken('much', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -757,7 +703,7 @@ function handleMuchLoop(parseContext)
  */
 function handleMany(parseContext)
 {
-  expectToken('many', parseContext);
+  tokenUtils.expectToken('many', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -822,7 +768,7 @@ function functionInvocation(functionName, parseContext, allowedSyntax)
  */
 function handlePlz(parseContext)
 {
-  expectAnyToken(['plz', '.plz'], parseContext);
+  tokenUtils.expectAnyToken(['plz', '.plz'], parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -931,7 +877,7 @@ function handleIncrementDecrement(parseContext)
  *  var <variable> = <value>;
  */
 function handleVery(parseContext) {
-  expectToken('very', parseContext);
+  tokenUtils.expectToken('very', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -994,7 +940,7 @@ function handleVery(parseContext) {
  */
 function handleLevl(parseContext)
 {
-  expectToken('levl', parseContext);
+  tokenUtils.expectToken('levl', parseContext);
 
   var tokens = parseContext.tokens;
 
@@ -1017,7 +963,7 @@ function handleLevl(parseContext)
  *  return <expression>;
  */
 function handleAmaze(parseContext) {
-  expectToken('amaze', parseContext);
+  tokenUtils.expectToken('amaze', parseContext);
 
   // consume: amaze
   parseContext.tokens.shift();
@@ -1085,7 +1031,7 @@ function parseStatement(parseContext) {
   // not dogescript, such javascript
   if ( !isDogescriptSource(parseContext) )
   {
-    return joinTokens(tokens) + '\n';
+    return tokenUtils.joinTokens(tokens) + '\n';
   }
 
   // trained use strict
@@ -1272,7 +1218,7 @@ function parseStatement(parseContext) {
   }
 
   var statement = tokens.shift();
-  // if there's more tokens, split them by a ' ' similar to how joinTokens would work
+  // if there's more tokens, split them by a ' ' similar to how tokenUtils.joinTokens would work
   if(tokens.length > 0)
   {
     statement += ' ';
